deepnsm: resolve TD-DEEPNSM-CLIPPY-195 — hand-reviewed clippy sweep
`cargo clippy --manifest-path crates/deepnsm/Cargo.toml --all-targets
-- -D warnings` is now clean (exit 0). Cleared 22 lints across 13 files:
the original 7-file set tracked as TD-DEEPNSM-CLIPPY-195 plus the lints
in this PR's new reader modules (window / reader_state /
crystal_neighborhood / sentence_transformer64 / signed_crystal /
codebook) that --all-targets surfaces.

All fixes are hand-applied and reviewed (NOT `clippy --fix`, which
mangled reader_state.rs into stranded-comment match guards):

- needless_range_loop: VsaVec::{bind,complement} -> std::array::from_fn;
  bundle outer loop + similarity tables -> iter_mut().enumerate().
- len_without_is_empty: is_empty() on Codebook + Crystal4096Neighbourhood.
- collapsible_if: reader_state Relative/Anaphora arms -> match guards with
  the explanatory comment moved above the arm (readable, unlike --fix).
- let_and_return, manual_contains, manual_(!)range_contains,
  manual_is_multiple_of, unnecessary_cast (x4), redundant_field_names,
  repeat().take()->repeat_n (x2), doc_lazy_continuation.
- Cargo.toml: move misplaced `description` from [workspace]
  (unused-manifest-key warning) to [package].

Pure idiom cleanup, no behavioural change. 217 tests green; fmt clean.

+++ crates/deepnsm/src/parser.rs
@@ -533,7 +533,7 @@ impl Parser {
                     resolved.push(r);
                     // Use the curated NSM-prime ID set rather than the
                     // earlier `r < 64` heuristic. See nsm_primes.rs.
-                    if crate::nsm_primes::is_nsm_prime(r as u16) {
+                    if crate::nsm_primes::is_nsm_prime(r) {
                         primes = primes.saturating_add(1);
                     }
                 }
